Mostly finished VR room, updated registration, added tables, etc.
Updated the reg form, now it creates a table for the user which contains their VR room preferences.

Rewrote a couple of functions for better security and to make breaking the site harder

Added more models, placed them in VR environment in the room

VR room features have been mostly finished. Still want to add a screen to visit friends and a button to go back home. Also one to leave the room. Will do so for the next version.